requirements.txt
streamlit
google-generativeai
langchain
langchain-community
sentence-transformers
pdf2image
Pillow
chromadb
base64
